refactor: 유저 신고 모델 공통 모델로 확장 [TOKTOT-424]
